+++ projects/fusion-ui/components/multi-dropdown/v1/multi-dropdown.module.ts
@@ -7,8 +7,8 @@ import {InputModule} from '@ironsource/fusion-ui/components/input/v1';
 import {FlagModule} from '@ironsource/fusion-ui/components/flag/v1';
 import {CheckboxModule} from '@ironsource/fusion-ui/components/checkbox/v1';
 import {ClickOutsideModule} from '@ironsource/fusion-ui/directives/click-outside';
-import {FilterByFieldModule} from '@ironsource/fusion-ui/pipes/collection';
-import {CloneModule} from '@ironsource/fusion-ui/pipes/clone';
+import {FilterByFieldPipe} from '@ironsource/fusion-ui/pipes/collection';
+import {ClonePipe} from '@ironsource/fusion-ui/pipes/clone';
 import {TooltipModule} from '@ironsource/fusion-ui/components/tooltip/v1';
 import {DynamicComponentsModule} from '@ironsource/fusion-ui/components/dynamic-components/v1';
 import {DropdownLoaderModule} from '@ironsource/fusion-ui/components/dropdown-loader/v1';
@@ -29,8 +29,8 @@ import {MultiDropdownComponent} from './multi-dropdown.component';
         FlagModule,
         CheckboxModule,
         ClickOutsideModule,
-        FilterByFieldModule,
-        CloneModule,
+        FilterByFieldPipe,
+        ClonePipe,
         TooltipModule,
         DynamicComponentsModule,
         DropdownLoaderModule,