+++ projects/fusion-ui/components/table/v1/table.module.ts
@@ -27,9 +27,9 @@ import {
 import {MultiDropdownModule} from '@ironsource/fusion-ui/components/multi-dropdown/v1';
 import {InputInlineModule} from '@ironsource/fusion-ui/components/input-inline/v1';
 import {LoadMoreModule} from '@ironsource/fusion-ui/directives/load-more';
-import {GenericPipeModule} from '@ironsource/fusion-ui/pipes/generic';
+import {GenericPipe} from '@ironsource/fusion-ui/pipes/generic';
 import {ErrorMessageModule} from '@ironsource/fusion-ui/components/error-message/v1';
-import {NotAvailableModule} from '@ironsource/fusion-ui/pipes/not-available';
+import {NotAvailablePipe} from '@ironsource/fusion-ui/pipes/not-available';
 import {ButtonModule} from '@ironsource/fusion-ui/components/button/v1';
 
 @NgModule({
@@ -59,10 +59,10 @@ import {ButtonModule} from '@ironsource/fusion-ui/components/button/v1';
         ToggleModule,
         ClickOutsideModule,
         ReactiveFormsModule,
-        GenericPipeModule,
+        GenericPipe,
         LoadMoreModule,
         ErrorMessageModule,
-        NotAvailableModule,
+        NotAvailablePipe,
         ButtonModule
     ]
 })