+++ projects/fusion-ui/components/table/v2/components/table-basic/table-basic.component.spec.ts
@@ -20,9 +20,9 @@ import {TableGroupedComponent} from '../table-grouped/table-grouped.component';
 import {TableLoadingComponent} from '../table-loading/table-loading.component';
 import {TableRowComponent} from '../table-row/table-row.component';
 import {TableRowGroupedComponent} from '../table-row-grouped/table-row-grouped.component';
-import {GenericPipeModule} from '@ironsource/fusion-ui/pipes/generic';
+import {GenericPipe} from '@ironsource/fusion-ui/pipes/generic';
 import {LoadMoreModule} from '@ironsource/fusion-ui/directives/load-more';
-import {NotAvailableModule} from '@ironsource/fusion-ui/pipes/not-available';
+import {NotAvailablePipe} from '@ironsource/fusion-ui/pipes/not-available';
 
 describe('TableBasicComponent', () => {
     let component: TableBasicComponent;
@@ -44,9 +44,9 @@ describe('TableBasicComponent', () => {
                 ToggleModule,
                 ClickOutsideModule,
                 ReactiveFormsModule,
-                GenericPipeModule,
+                GenericPipe,
                 LoadMoreModule,
-                NotAvailableModule
+                NotAvailablePipe
             ],
             declarations: [
                 TableBasicComponent,