+++ projects/fusion-ui/pipes/not-available/public-api.ts
@@ -1,2 +1,1 @@
-export * from './not-available.module';
 export * from './not-available.pipe';