+++ projects/fusion-ui/services/api/api.module.ts
@@ -1,9 +1,0 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {CapitalizeModule} from '@ironsource/fusion-ui/pipes/string';
-
-@NgModule({
-    declarations: [],
-    imports: [CommonModule, CapitalizeModule]
-})
-export class ApiModule {}