+++ CHANGELOG.md
@@ -1,8 +1,37 @@
 # Changelog
 
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
-## [4.0.0](https://github.com/ironSource/fusion-ui/compare/v4.0.0-rc.8...v4.0.0) (2022-10-03)
+### [5.0.1](https://github.com/ironSource/fusion-ui/compare/v5.0.1-rc.2...v5.0.1) (2022-11-07)
+
+### [5.0.1-rc.2](https://github.com/ironSource/fusion-ui/compare/v5.0.1-rc.1...v5.0.1-rc.2) (2022-11-07)
+
+### [5.0.1-rc.1](https://github.com/ironSource/fusion-ui/compare/v4.1.0...v5.0.1-rc.1) (2022-11-06)
+
+### Code Refactoring
+
+* **fu-479:** capitalize pipe to standalone with providedIn:root ([0686271](https://github.com/ironSource/fusion-ui/commit/0686271faea6596fe2275e50ef3959640fa775aa))
+* **fu-479:** clone pipes to standalone ([5ae0161](https://github.com/ironSource/fusion-ui/commit/5ae016160b32a40ae04bcd5537cc10767a74e27b))
+* **fu-479:** collection pipes to standalone ([1e0cdca](https://github.com/ironSource/fusion-ui/commit/1e0cdcaae9b7936b69be77f4b37136122e74605c))
+* **fu-479:** generic pipe to standalone ([4445483](https://github.com/ironSource/fusion-ui/commit/444548321226f3a15d5f238918be8240c678d023))
+* **fu-479:** notAvailable pipe to standalone ([5b846a0](https://github.com/ironSource/fusion-ui/commit/5b846a0c535aa69c038d58bac25f23e1203ae67e))
+* **fu-479:** secondsToMinutes and shortNumberScaleSuffix pipes to standalone ([f8024a0](https://github.com/ironSource/fusion-ui/commit/f8024a0228216b4f9c27a2426be57655cd055bc2))
+* **fu-479:** truncate pipe to standalone ([a2410bb](https://github.com/ironSource/fusion-ui/commit/a2410bb9ba97c66e18c2e38be169b28cde71fd2e))
+
+### CI
+
+* **fu-479:** switch from standard-version to commit-and-tag-version ([29600de](https://github.com/ironSource/fusion-ui/commit/29600ded5b371220badefe4d92e2144e15aa21a5))
+
+## [4.1.0](https://github.com/ironSource/fusion-ui/compare/v4.0.0...v4.1.0) (2022-10-26)
+
+### Features
+
+* **4.x.x:** fix modal closing ([ffb56b0](https://github.com/ironSource/fusion-ui/commit/ffb56b008cd033752734a0eb69642923b4b20af9))
+
+## [4.0.0](https://github.com/ironSource/fusion-ui/compare/v3.6.0...v4.0.0) (2022-10-03)
+
+## [4.0.0-rc.3](https://github.com/ironSource/fusion-ui/compare/v3.4.0-rc.11...v4.0.0-rc.3) (2022-09-19)
+
 
 ### Bug Fixes
 
@@ -18,12 +47,19 @@ All notable changes to this project will be documented in this file. See [standa
 
 * **fu-461:** dependencies update ([3c9ffc4](https://github.com/ironSource/fusion-ui/commit/3c9ffc4d577f2d142c64c8c88f886b655bf27ad3))
 
+## [4.0.0-rc.1](https://github.com/ironSource/fusion-ui/compare/v3.3.0...v4.0.0-rc.1) (2022-09-06)
+
+
+### Bug Fixes
+
+* **fu-461:** dependencies update ([3c9ffc4](https://github.com/ironSource/fusion-ui/commit/3c9ffc4d577f2d142c64c8c88f886b655bf27ad3))
+
 ## [4.0.0-rc.0](https://github.com/ironSource/fusion-ui/compare/v3.2.4...v4.0.0-rc.0) (2022-09-01)
 
 
 ### ⚠ BREAKING CHANGES
 
-* **fu-457:** angular version upgraded to v14
+* **fu-457:** upgrade angular to angular 14
 
 ### Features
 
@@ -33,6 +69,20 @@ All notable changes to this project will be documented in this file. See [standa
 ### CI
 
 * **upgrade:** update sass, sass-loader and sass-bundle ([a9cbe07](https://github.com/ironSource/fusion-ui/commit/a9cbe07c35d21fa697485da46c722e178d82ef3d))
+
+### [3.7.1](https://github.com/ironSource/fusion-ui/compare/v3.7.1-rc.0...v3.7.1) (2022-10-20)
+
+### [3.7.1-rc.0](https://github.com/ironSource/fusion-ui/compare/v3.7.0...v3.7.1-rc.0) (2022-10-19)
+
+
+### Bug Fixes
+
+* **isModalOpen:** fix modal closing ([ffb56b0](https://github.com/ironSource/fusion-ui/commit/ffb56b008cd033752734a0eb69642923b4b20af9))
+
+
+### CI
+
+* **fu-474:** change main ci workflow ([c7ee466](https://github.com/ironSource/fusion-ui/commit/c7ee466790c94313aafc6450a99516c49b8e4601))
 
 ## [3.7.0](https://github.com/ironSource/fusion-ui/compare/v3.7.0-rc.1...v3.7.0) (2022-10-12)